Add function to export pdf files
Now users can download a pdf with their current lists and tasks.
This function is located into the navbar. Is visible when the user hover
his name.
To made this, I created an 'exportable' controller, here I wrote the
'export' method. In this controller I'll make a function to export CSV
files with the same data.
By the way. The main toot that I used is Prawn. It's a very useful gem
to generate PDF's