Add export CSV function
I re-used the 'exportable' controller, to switch between pdf or csv
export. I added a function into List model. This function generate a
CSV file with all current user  tasks.

The user can run this function with a button added into the navbar.
To see it, user must hover his name. There is a dropdown with
a group of buttons